+++ blueprints/names/tests/test_hello_name.py
@@ -4,7 +4,7 @@ from flask import url_for
 
 @pytest.fixture
 def resp(client):
-    response = client.get(url_for('hello_name', name='Renzo'))
+    response = client.get(url_for('names.hello_name', name='Renzo'))
     return response
