+++ blueprints/names/tests/test_query_string.py
@@ -5,7 +5,7 @@ from flask import url_for
 @pytest.fixture
 def resp(client):
     dct = {'name': 'Renzo', 'lastname': 'Nuccitelli'}
-    return client.get(url_for('name_lastname'), query_string=dct)
+    return client.get(url_for('names.name_lastname'), query_string=dct)
 
 
 def test_status_code(resp):
