--- rotas/hello.py
@@ -1,11 +1,0 @@
-from app import app
-
-
-@app.route('/')
-def hello():
-    return 'Olá Mundo'
-
-
-@app.route('/nome/<name>')
-def hello_name(name):
-    return f'Olá {name}'
